cleaning up and fixing tests
