--- include/bindings/tlstm_layer_bindings.h
@@ -1,6 +1,0 @@
-#pragma once
-#include <pybind11/pybind11.h>
-
-#include "../tlstm_layer.h"
-
-void bind_tlstm_layer(pybind11::module_& modo);
