+++ include/lstm_layer_cuda.cuh
@@ -9,52 +9,41 @@
 class LSTMCuda : public BaseLayerCuda {
    public:
     int _batch_size = -1;
-    float act_omega = 0.0000001f;
     float gain_w;
     float gain_b;
     std::string init_method;
     int w_pos_f, b_pos_f, w_pos_i, b_pos_i, w_pos_c, b_pos_c, w_pos_o, b_pos_o;
+    bool last_timestep = false;
 
     LSTMStateCuda lstm_state;
 
-    LSTMCuda(size_t input_size, size_t output_size, int seq_len,
-             bool bias = true, float gain_w = 1.0f, float gain_b = 1.0f,
-             std::string init_method = "Xavier", int device_idx = 0);
+    // Backward temp buffers (cached across calls)
+    float *d_buf_rec_mu = nullptr, *d_buf_rec_var = nullptr;
+    float *d_buf_combined_mu = nullptr, *d_buf_combined_var = nullptr;
+    float *d_buf_xh_mu = nullptr, *d_buf_xh_var = nullptr;
+    float *d_buf_sum_w = nullptr, *d_buf_sum_b = nullptr;
+
+    LSTMCuda(size_t input_size, size_t output_size, bool last_timestep = false,
+             int seq_len = 1, bool bias = true, float gain_w = 1.0f,
+             float gain_b = 1.0f, std::string init_method = "Xavier",
+             int device_idx = 0);
 
     ~LSTMCuda();
 
-    // NOTE: Make the class movable not copyable
-    // Delete copy constructor and copy assignment
     LSTMCuda(const LSTMCuda &) = delete;
     LSTMCuda &operator=(const LSTMCuda &) = delete;
-
-    // Optionally implement move constructor and move assignment
     LSTMCuda(LSTMCuda &&) = default;
     LSTMCuda &operator=(LSTMCuda &&) = default;
 
     std::string get_layer_info() const override;
-
     std::string get_layer_name() const override;
-
     LayerType get_layer_type() const override;
-
     int get_input_size() override;
-
     int get_output_size() override;
+    int get_max_num_states() override;
 
     void get_number_param();
-
     void init_weight_bias() override;
-
-    void prepare_input(BaseHiddenStates &input_state);
-
-    void forget_gate(int batch_size);
-
-    void input_gate(int batch_size);
-
-    void cell_state_gate(int batch_size);
-
-    void output_gate(int batch_size);
 
     void forward(BaseHiddenStates &input_states,
                  BaseHiddenStates &output_states,
@@ -66,7 +55,7 @@ class LSTMCuda : public BaseLayerCuda {
                   bool state_udapte = true) override;
 
     std::unique_ptr<BaseLayer> to_host() override;
-
+    void to(int device_idx) override;
     void preinit_layer() override;
 
     void d_get_LSTM_states(std::vector<float> &mu_h, std::vector<float> &var_h,
@@ -77,9 +66,12 @@ class LSTMCuda : public BaseLayerCuda {
                            const std::vector<float> &var_h,
                            const std::vector<float> &mu_c,
                            const std::vector<float> &var_c);
-    void to(int device_idx) override;
 
    protected:
     using BaseLayerCuda::allocate_param_memory;
     using BaseLayerCuda::params_to_device;
+
+   private:
+    void allocate_bwd_buffers(int batch_size);
+    void deallocate_bwd_buffers();
 };
