+++ include/slstm_layer.h
@@ -9,12 +9,12 @@ class SLSTM : public LSTM {
    public:
     SmoothSLSTM smooth_states;
     int time_step = 0;
-    SLSTM(size_t input_size, size_t output_size, int seq_len = 1,
-          bool bias = true, float gain_w = 1.0f, float gain_b = 1.0f,
-          std::string init_method = "Xavier", int time_step = 0,
-          int device_idx = 0)
-        : LSTM(input_size, output_size, seq_len, bias, gain_w, gain_b,
-               init_method, device_idx),
+    SLSTM(size_t input_size, size_t output_size, bool last_timestep = false,
+          int seq_len = 1, bool bias = true, float gain_w = 1.0f,
+          float gain_b = 1.0f, std::string init_method = "Xavier",
+          int time_step = 0, int device_idx = 0)
+        : LSTM(input_size, output_size, last_timestep, seq_len, bias, gain_w,
+               gain_b, init_method, device_idx),
           time_step(time_step) {}
 
     std::string get_layer_info() const override;
@@ -24,6 +24,11 @@ class SLSTM : public LSTM {
     LayerType get_layer_type() const override;
 
     void prepare_input_smooth(SmoothingHiddenStates &input_state);
+
+    void forget_gate(int batch_size);
+    void input_gate(int batch_size);
+    void cell_state_gate(int batch_size);
+    void output_gate(int batch_size);
 
     void forward(BaseHiddenStates &input_states,
                  BaseHiddenStates &output_states,
