+++ src/bindings/lstm_layer_bindings.cpp
@@ -5,15 +5,13 @@
 
 #include "../include/lstm_layer.h"
 
-void bind_lstm_layer(pybind11::module_& modo)
-/*
- */
-{
+void bind_lstm_layer(pybind11::module_& modo) {
     pybind11::class_<LSTM, std::shared_ptr<LSTM>, BaseLayer>(modo, "LSTM")
-        .def(pybind11::init<size_t, size_t, int, bool, float, float,
+        .def(pybind11::init<size_t, size_t, bool, int, bool, float, float,
                             std::string, int>(),
              pybind11::arg("input_size"), pybind11::arg("output_size"),
-             pybind11::arg("seq_len"), pybind11::arg("bias"),
+             pybind11::arg("last_timestep") = false,
+             pybind11::arg("seq_len") = 1, pybind11::arg("bias") = true,
              pybind11::arg("gain_weight") = 1.0f,
              pybind11::arg("gain_bias") = 1.0f, pybind11::arg("method") = "He",
              pybind11::arg("device_idx") = 0)
