+++ test/py_unit/test_attention.py
@@ -34,7 +34,7 @@ class TestMultiheadAttention(unittest.TestCase):
         )
 
         np.random.seed(42)
-        input_data = np.random.randn(batch_size * seq_len * embed_dim).astype(
+        input_data = np.random.randn(batch_size, seq_len, embed_dim).astype(
             np.float32
         )
